Moved loading data into its own file and added minst dataset to test with
